Don't throw frivolous exceptions
Previously, an exception was thrown and caught every time the player dies. This is bad for performance and messes with debugging efforts that involve breaking when a handled exception is thrown.
With this change, system behavior does not change in an observable way, but an exception is not thrown in the process.